util: move pod annotation code into its own file, add tests
